pyproject.toml
[tool.poetry]
name = "app-blog"
version = "0.1.0"
description = ""
authors = ["c1kvoy <seva33039@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.32"
greenlet = "^3.0.3"
uvicorn = "^0.30.4"
pydantic-settings = "^2.4.0"
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
cryptography = "^43.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
